examples: Don't expect all domains to have a disk
When generating the list of domains inside of index.php the
get_disk_stats() function from libvirt.php is called. Now, if the
domain has no disks a warning is generated and subsequent loop
produces another warning (because libvirt_domain_get_block_info()
is expected to return an array or false in case of failure).

Silent the call to libvirt_domain_get_block_info() and execute
the loop iff an actual array was returned.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>