examples: Force int division in translate_perms()
The translate_perms() function (implemented in libvirt.php) is
responsible for translating an octal representation of perms into
a string. To do this, it looks at the very last digit of the
mode, translates it and then divides the mode by 10 to shift the
mode right. Well, '$mode /= 10;' in PHP means floating point
division. To enforce integer division intdiv() has to be used
instead.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>