examples: Fix printing storage pool table
Firstly, there are few discrepancies in closing tags (in fact,
they lack the closing '/' and thus open another tag).
Secondly, the nested table of storage volumes is supposed to span
over all columns of the parent table (storage pool). Well, it
doesn't as it spans only over 10 columns while the parent table
has 11 columns. While at it, also set the style of the nested
table to match the width of the parent table.
And finally, put new line character into various echo statements
so that the page source is easier to read.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>