examples: Don't expect all storage pools to have Linux permissions
Some types of storage pools (typically an inactive iSCSI) do not
report linux permissions in their XMLs. Unfortunately, the way
libvirt_storagepool_get_xml_desc() is designed is that when given
XPATH returns no nodes (or fails to evaluate) then the original
XML is returned. This leaves caller with no easy way to check
whether XPATH was evaluated or not. But try anyway.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>